refactor: consolidate testing utils
