+++ banking/testing_utils.py
@@ -1,0 +1,25 @@
+# Copyright (c) 2025, ALYF GmbH and Contributors
+# See license.txt
+
+import frappe
+
+TEST_COMPANY = "Bolt Trades"
+
+
+def create_currency_account(currency: str, parent_account: str, account_name: str):
+	acc = frappe.new_doc("Account")
+	acc.account_name = account_name
+	acc.account_currency = currency
+	acc.parent_account = parent_account
+	acc.insert(ignore_permissions=True, ignore_mandatory=True, ignore_links=True)
+	return acc
+
+
+def create_bank_account(account: str):
+	ba = frappe.new_doc("Bank Account")
+	ba.account_name = "_Test_B_Account"
+	ba.account = account
+	ba.bank = "_Test_Bank"
+	ba.is_company_account = 1
+	ba.insert(ignore_permissions=True, ignore_links=True)
+	return ba
